+++ src/heuristics/partialOrder/PartialOrderHeuristicGiven.hpp
@@ -30,7 +30,7 @@ namespace d4 {
 
 class PartialOrderHeuristicGiven : public PartialOrderHeuristic {
  protected:
-  std::vector<unsigned> m_order;
+  std::vector<double> m_order;
   double m_scaleFactor;
 
  public:
@@ -57,7 +57,7 @@ class PartialOrderHeuristicGiven : public PartialOrderHeuristic {
    *
    * @return the position of v in the order.
    */
-  inline unsigned getPartialOrder(Var v) override {
+  inline double getPartialOrder(Var v) override {
     return m_order[v];
   }  // getPartialOrder
